docs(issues): remove support for focus-visible polyfill #330
Remove the dependency from the wc-generator and update the documentation to support this transition. Retrofitting legacy elements is not a priority as there will be a time overlap while the population upgrades to the next version.

Changes to be committed:
modified: template/package.json
modified: docs/readme.md
deleted: docs/focus-visible.md