perf(docs): remove comments removal script
This commit removes a previous process that would
remove all comments from generated files. This
was discussed and proved to be problematic when working
with generated documentation. Users were confused by the
lack of comments that eluded to them being generated.

Changes to be committed:
modified:   template/scripts/generateDocs.js