fix: update inefficient RegEx #357
- Update below regex to make it more efficient:
\r\n|\r|\n -> \r?\n|\r(?!\n)

1. The \r\n is redundant to \r checking in current pattern,
and I think we can combine it with the \n checking by using \r?\n
2. At first, I want to also remove the standalone \r because it only for older Mac OS,
but I think we should keep it for backward compatibility, so adding | \r to the pattern.
3. Additional (?!\n) in the end is to make sure the \r is not followed by \n,
to avoid exponential backtracking on strings starting with '''
and containing many repetitions of '\r\n'. (per latest [CodeQL analysis alert](https://github.com/AlaskaAirlines/WC-Generator/security/code-scanning/3))

Changes to be committed:
	modified:   package-lock.json
	modified:   template/scripts/generateDocs.mjs